chore: organize auth components in dedicated folder
